If setup does not find zp-config.php it now copies zp-config.php.example to zp-config.php.
New instructions for install. Just copy the zenphoto install files to the server and run setup. When prompted, enter the information for the database:
 
mySQL Admin user name
mySQL Admin password
mySQL host (not normally necessary to change the default: localhost)
mySQL database name for the zenphoto album
Database prefix for the zenphoto tables (may be left blank if the database is not shared)

git-svn-id: http://www.zenphoto.org/svn/trunk@940 46129fc4-274e-7f48-8cfa-5da4c607681c