sanitize function can now automatically detect arrays and process them as such.
theme definition text is now filtered by KSES whether the text is provided in a .txt or .php file.

all calls to the sanitize function have been converted to use the new sanitize function based on the context.

direct calls to the kses library have been eliminated.  this should make it easier to upgrade to a different filter in the future if necessary.

git-svn-id: http://www.zenphoto.org/svn/trunk@2117 46129fc4-274e-7f48-8cfa-5da4c607681c